requirements.txt
boto3
click
prompt_toolkit
psycopg2
flask
infi.systray
jaraco.clipboard
ahk
typer
PyGithub
Pillow